refactor(web-app): Simplified navlinks
Moved auth buttons to auth module and pushed background-changing logic up to parent
