+++ src/Layouts/app/AppLayout.tsx
@@ -1,9 +1,8 @@
 /* eslint-disable import/no-unresolved */
 import { DeveloperConsole } from 'chocolate/substrate-lib/components';
-/* eslint-enable import/no-unresolved */
-import { useState } from 'react';
 import { Toaster } from 'react-hot-toast';
 import { ReactQueryDevtools } from 'react-query/devtools';
+import { useMatch } from 'react-router-dom';
 import MenuBar from '../../customComponents/menuBar';
 import { loader, message } from '../../customComponents/utilities';
 import { useSubstrate } from '../../substrate-lib';
@@ -12,14 +11,13 @@ import { AppRoutes } from './components/AppRoutes';
 /* NB: AppLayout==AppShell, so do  all box styling here */
 export function AppLayout(): JSX.Element {
   const { apiState, apiError } = useSubstrate();
-  const [back, setBack] = useState(false);
+  const match = useMatch('/');
 
   if (apiState === 'ERROR') return message(apiError);
   if (apiState !== 'READY') return loader('Connecting to Substrate');
-
   return (
-    <div className={`root-wrap ${back ? 'background' : ''}`}>
-      <MenuBar setBack={setBack} />
+    <div className={`root-wrap ${match ? 'background' : ''}`}>
+      <MenuBar />
       <AppRoutes />
       <Toaster position='bottom-right' />
       {process.env.NODE_ENV === 'development' && <ReactQueryDevtools />}
